feat: create components module and move toggleButton.tsx here(not working version)
