+++ components/buttons/README.md
@@ -1,0 +1,12 @@
+# Buttons Package
+
+This package contains buttons components.
+To use the code in the package, you need to add the following to the package.json file.
+
+```json
+{
+  "dependencies": {
+    "@chrome-extension-boilerplate/buttons": "workspace:*"
+  }
+}
+```
