+++ components/buttons/lib/toggleButton.tsx
@@ -1,0 +1,18 @@
+import React, { ComponentPropsWithoutRef } from 'react';
+import { useStorageSuspense } from '@chrome-extension-boilerplate/shared';
+import { exampleThemeStorage } from '@chrome-extension-boilerplate/storage';
+
+export const ToggleButton = (props: ComponentPropsWithoutRef<'button'>) => {
+  const theme = useStorageSuspense(exampleThemeStorage);
+  return (
+    <button
+      className={
+        props.className +
+        'font-bold mt-4 py-1 px-4 rounded shadow hover:scale-105 ' +
+        (theme === 'light' ? 'bg-white text-black' : 'bg-black text-white')
+      }
+      onClick={exampleThemeStorage.toggle}>
+      {props.children}
+    </button>
+  );
+};
