+++ pages/options/src/Options.tsx
@@ -1,7 +1,7 @@
 import '@src/Options.css';
 import { useStorageSuspense, withErrorBoundary, withSuspense } from '@chrome-extension-boilerplate/shared';
 import { exampleThemeStorage } from '@chrome-extension-boilerplate/storage';
-import { ComponentPropsWithoutRef } from 'react';
+import { ToggleButton } from '@chrome-extension-boilerplate/buttons';
 
 const Options = () => {
   const theme = useStorageSuspense(exampleThemeStorage);
@@ -17,22 +17,6 @@ const Options = () => {
       Edit <code>pages/options/src/Options.tsx</code> and save to reload.
       <ToggleButton>Toggle theme</ToggleButton>
     </div>
-  );
-};
-
-const ToggleButton = (props: ComponentPropsWithoutRef<'button'>) => {
-  const theme = useStorageSuspense(exampleThemeStorage);
-  return (
-    <button
-      className={
-        props.className +
-        ' ' +
-        'font-bold mt-4 py-1 px-4 rounded shadow hover:scale-105 ' +
-        (theme === 'light' ? 'bg-white text-black' : 'bg-black text-white')
-      }
-      onClick={exampleThemeStorage.toggle}>
-      {props.children}
-    </button>
   );
 };
