+++ pnpm-workspace.yaml
@@ -2,3 +2,4 @@ packages:
   - "chrome-extension"
   - "pages/*"
   - "packages/*"
+  - "components/*"
